refactor(quality): rename DatasetQualityItem to QualityItem
The component is shared between datasets and dataservices.
