+++ datagouv-components/src/components/QualityItem.vue
